Move mailbox related code from P2PService to MailboxMessageService
